Add JSON tags, improve formatting of debug locations, use camel case for url querys, add appID to wishes and sepparate middlewares implementation
